+++ src/main/java/com/example/service/IEmpleadoService.java
@@ -7,9 +7,9 @@ import com.example.entities.Empleado;
 public interface IEmpleadoService {
 
     public List<Empleado> getEmpleados();
-    public Empleado getEmpleado(long idEmpleado);
+    public Empleado getEmpleado(int idEmpleado);
     public void guardar(Empleado empleado);
-    public void eliminar(long idEmpleado);
-    public void modificar(long idEmpleado, Empleado empleado);
+    public void eliminar(int idEmpleado);
+    public void modificar(int idEmpleado, Empleado empleado);
     
 }
